Do not change the protocol representation

+++ commons/zenoh-protocol/src/core/encoding.rs
@@ -16,7 +16,7 @@ use alloc::borrow::Cow;
 use core::fmt::{self, Debug};
 use zenoh_result::{bail, ZResult};
 
-pub type EncodingPrefix = u64;
+pub type EncodingPrefix = u8;
 
 /// The encoding of a zenoh `zenoh::Value`.
 ///
